A nice little restructure that let's use split the dockerfiles and the jenkins files
