+++ inserts/.gitignore
@@ -1,0 +1,5 @@
+# IntelliJ project files
+.idea
+*.iml
+*.json
+/GLOBALS.py
