+++ inserts/kafka_kino_inserts.py
@@ -1,9 +1,9 @@
 import json
 import os
 import sys
-import GLOBALS
-
 from kafka import KafkaConsumer, KafkaProducer
+
+from apis import GLOBALS
 
 try:
     table_name = __import__("insert_{}".format(os.environ['TABLE_NAME']))
@@ -18,7 +18,7 @@ try:
     KAFKA_BROKER = os.environ['KAFKA_BROKER']
 except KeyError:
     try:
-        from GLOBALS import KAFKA_BROKER
+        from apis.GLOBALS import KAFKA_BROKER
     except ImportError:
         print("Specify Kafka Brokers")
         exit()
