+++ inserts/tests/test_insert_movies2numbers.py
@@ -1,9 +1,9 @@
-import unittest
 import json
-import GLOBALS
+import unittest
 
-from postgres import Postgres
-from insert_movies2genres import InsertMovies2Genres
+from apis import GLOBALS
+from inserts.insert_movies2numbers import InsertMovies2Numbers
+from py.postgres import Postgres
 
 server = GLOBALS.PG_SERVER
 port = GLOBALS.PG_PORT
@@ -14,11 +14,11 @@ pw = GLOBALS.PG_PASSWORD
 with open('test_data.json') as data_file:
     data = json.load(data_file)
 
-class TestInsertMovies2Genres(unittest.TestCase):
+class TestInsertMovies2Numbers(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.ins = InsertMovies2Genres(server, port, db, user, pw)
+        cls.ins = InsertMovies2Numbers(server, port, db, user, pw)
         cls.pg = Postgres(server, port, db, user, pw)
         # We insert the corresponding film into kino.movies
         # due to the foreign key constraint.
@@ -29,14 +29,15 @@ class TestInsertMovies2Genres(unittest.TestCase):
 
     def test_insert_movies(self):
         self.ins.insert(data)
-        self.pg.pg_cur.execute('select imdb_id, genre from kino.movies2genres')
+        self.pg.pg_cur.execute('select imdb_id, value, type from kino.movies2numbers')
         result = self.pg.pg_cur.fetchall()
-        self.assertEqual(result, [('tt2562232', 'Drama'), ('tt2562232', 'Comedy')])
+        self.assertEqual(result, [('tt2562232', 103215000, 'revenue'), ('tt2562232', 18000000, 'budget')])
+
 
     @classmethod
     def tearDownClass(cls):
         cls.pg = Postgres(server, port, db, user, pw)
-        cls.pg.pg_cur.execute('delete from kino.movies2genres')
+        cls.pg.pg_cur.execute('delete from kino.movies2numbers')
         cls.pg.pg_cur.execute('delete from kino.movies')
         cls.pg.pg_conn.commit()
