+++ inserts/tests/test_insert_movies2posters.py
@@ -1,9 +1,9 @@
-import unittest
 import json
-import GLOBALS
+import unittest
 
-from postgres import Postgres
-from insert_movies2keywords import InsertMovies2Keywords
+from apis import GLOBALS
+from inserts.insert_movies2posters import InsertMovies2Posters
+from py.postgres import Postgres
 
 server = GLOBALS.PG_SERVER
 port = GLOBALS.PG_PORT
@@ -14,11 +14,11 @@ pw = GLOBALS.PG_PASSWORD
 with open('test_data.json') as data_file:
     data = json.load(data_file)
 
-class TestInsertMovies2Keywords(unittest.TestCase):
+class TestInsertMovies2Posters(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.ins = InsertMovies2Keywords(server, port, db, user, pw)
+        cls.ins = InsertMovies2Posters(server, port, db, user, pw)
         cls.pg = Postgres(server, port, db, user, pw)
         # We insert the corresponding film into kino.movies
         # due to the foreign key constraint.
@@ -29,21 +29,14 @@ class TestInsertMovies2Keywords(unittest.TestCase):
 
     def test_insert_movies(self):
         self.ins.insert(data)
-        self.pg.pg_cur.execute('select imdb_id, keyword from kino.movies2keywords')
+        self.pg.pg_cur.execute('select imdb_id, url from kino.movies2posters')
         result = self.pg.pg_cur.fetchall()
-        self.assertEqual(result, [('tt2562232', 'times square'),
-                                  ('tt2562232', 'superhero'),
-                                  ('tt2562232', 'long take'),
-                                  ('tt2562232', 'new york city'),
-                                  ('tt2562232', 'play'),
-                                  ('tt2562232', 'broadway'),
-                                  ('tt2562232', 'actor')
-                                  ])
+        self.assertEqual(result, [('tt2562232', 'http://image.tmdb.org/t/p/w185//rSZs93P0LLxqlVEbI001UKoeCQC.jpg')])
 
     @classmethod
     def tearDownClass(cls):
         cls.pg = Postgres(server, port, db, user, pw)
-        cls.pg.pg_cur.execute('delete from kino.movies2keywords')
+        cls.pg.pg_cur.execute('delete from kino.movies2posters')
         cls.pg.pg_cur.execute('delete from kino.movies')
         cls.pg.pg_conn.commit()
